Develop controller for Cupom, Produto and Carrinho

+++ facilit-ecommerce/src/main/resources/db/migration/V002__cria-tabela-produto.sql
@@ -1,6 +1,6 @@
 create table produto(
-  id    integer auto_increment,
-  nome  varchar(100),
-  preco decimal(12,2),
+  id    	bigint auto_increment,
+  nome  	varchar(100),
+  preco 	decimal(12,2),
   primary key (id)
 );
